Updated robo_cylinder node
