+++ robo_cylinder/srv/StringCmd.srv
@@ -1,0 +1,3 @@
+string command
+---
+bool response
